playeradapter: seeking switched from using int64_t to double
Only being able to seek in increments of whole seconds was becoming a
problem. The subtitle list needs precision less than seconds to
accurately seek.

+++ src/gui/mpvadapter.cpp
@@ -582,7 +582,7 @@ void MpvAdapter::addSubtitle(const QString &file)
     }
 }
 
-void MpvAdapter::seek(const int64_t time)
+void MpvAdapter::seek(const double time)
 {
     QByteArray timestr = QString::number(time).toUtf8();
     const char *args[4] = {
